Progrese en controllers y services
termine todo los de controllers y services, lo uncio que falta es terminar la funcion upload de ambos archivos

Ademas rtengo que probar que todas las funciones funcionen correctamenet y que no haya errores,

Despues tengo que integrar la ia de categorizacion a la funcion uploadedMovements

Y finalizado eso ya estaria la parte principal terminada

+++ controllers/movements.controller.js
@@ -3,7 +3,7 @@ import movementService from "../services/movements.service.js";
 
 //Devuelve todos los movimientos financieros atravez de userId.
 const getMovements = async (req, res) => {
-    const userId = req.params.userId;
+    const userId = req.id;
     
     if (!userId) {
         return res.status(400).json({message: "Se requiere userId."}); }
@@ -16,24 +16,50 @@ const getMovements = async (req, res) => {
     }  
 };
 
-//Devuelve un movimiento financiero atravez del id del movimiento financiero.
-const getMovement = async (req, res) => {
-    const Movemnetid = req.params.id;
 
-    if (!id) {
-        return res.status(400).json({message: "Se requiere id del movimiento financiero"}); }
+//Devuelve un movimiento financiero atravez del id del movimiento financiero.
+const getMovementsByUser = async (req, res) => {
+    const userId = req.id;
+
+    if (!userId) {
+        return res.status(400).json({message: "Se requiere userId"}); }
 
     try{
-        const result = await movementService.getMovement(id);
+        const result = await movementService.getMovementByUser(userId);
+        if (!result){
+            return res.status(404).json({message: "Movimiento no encontrado"});}
+
         res.status(200).json(result.rows);
     } catch(error) {
         res.status(500).json({message: error.message});
     }
 };
 
-//
+//Devuelve un movimiento financiero atravez del id del movimiento financiero.
+const getMovementById = async (req, res) => {
+    const id = req.params.id;
+
+    if (!id) {
+        return res.status(400).json({message: "Se requiere id del movimiento financiero"}); }
+
+    try{
+        const result = await movementService.getMovementById(id);
+        if (!result){
+            return res.status(404).json({message: "Movimiento no encontrado"});}
+
+        res.status(200).json(result.rows);
+    } catch(error) {
+        res.status(500).json({message: error.message});
+    }
+};
+
+/*
+
+Falta hacer e upload Movements, terminarlo a logica del controller y de los services
+
+*/
 const uploadMovements = async (req, res) => {
-    const userId = req.params.userId;
+    const userId = req.id;
     const movements_exel = req.body;
 
     if (!userId){
@@ -52,7 +78,8 @@ const uploadMovements = async (req, res) => {
 
 //Crea un solo movimiento financiero
 const createMovement = async (req, res) => {
-    const { userId, fecha, categoria, monto, descripcion } = req.body;
+    const userId = req.id;
+    const { fecha, categoria, monto, descripcion } = req.body;
 
     if (!userId){
         return res.status(400).json({message: "Se necesita un userId"}); }
@@ -71,9 +98,9 @@ const createMovement = async (req, res) => {
 
     try{
         await movementService.createMovement(userId, fecha, categoria, monto, descripcion);
-        res.status(200).json({message: "Movimiento creado correctamente."});
+        res.status(201).json({message: "Movimiento creado correctamente."});
     } catch(error){
-        res.status(500).json({message: error.message});
+        res.status(500).json({message: "Error al crear el Movmiento", error});
     }
 };
 
@@ -93,6 +120,12 @@ const updateMovement = async (req, res) => {
     if (!categoria){
         return res.status(400).json({message: "Se necesita una categoria"}); }
 
+    //Ingresar cada una de las categorias posibles
+    /* 
+    if (categoria != "" || categoria != "" || categoria != "" || categoria != "" || categoria != "" ||categoria != "" || categoria != "" ){
+        return res.status(400).json({message: "Se necesita una categoria valida, $(categoria), no es valido"}); }
+    */
+
     if (!monto){
         return res.status(400).json({message: "Se necesita un monto"}); }
 
@@ -100,7 +133,14 @@ const updateMovement = async (req, res) => {
         return res.status(400).json({message: "Se necesita una descripcion"}); }
 
     try{
+        const movement = await movementService.getMovementById(id);
+
+        if (!movement) {
+            return res.status(404).json({ message: "movement no encontrado" }); 
+        }
+
         await movementService.updateMovement(id, fecha, categoria, monto, descripcion);
+
         res.status(200).json({message: "Movimiento actualizado correctamente."});
     } catch(error){
         res.status(500).json({message: error.message});
@@ -114,7 +154,14 @@ const deleteMovement = async (req, res) => {
         return res.status(400).json({message: "Se necesita el id del movmiento financiero"}); }
     
     try{
+        const movement = await movementService.getMovementById(id);
+
+        if (!movement) {
+            return res.status(404).json({ message: "movement no encontrado" }); 
+        }
+
         await movementService.deleteMovement(id);
+        
         res.status(200).json({message: "Movimiento eliminado correctamente."});
     } catch(error){
         res.status(500).json({message: error.message});
@@ -123,7 +170,8 @@ const deleteMovement = async (req, res) => {
 
 export default {
     getMovements,
-    getMovement,
+    getMovementsByUser,
+    getMovementById,
     uploadMovements,
     createMovement,
     updateMovement,
